TST: separate out groupby/test_nth
